Add Person class and fix culture-related issues
A new `Person` class was added in the `Packt.Shared` namespace, though it is currently empty. Fixed a typo in the `using` directive in `Program.Helpers.cs` (`Sytem.Globalization` corrected to `System.Globalization`). Adjusted formatting in a `WriteLine` statement to remove an extra space. In `Program.cs`, instantiated a `Person` object and wrote its `ToString` output to the console.